+++ test.sh
@@ -1,6 +1,12 @@
 #!/bin/bash -ex
 
 VERSION=$(< VERSION)
-docker run --rm debify:$VERSION config script > docker-debify
+
+ARCH="$1"
+if [ -z "$ARCH" ]; then
+  ARCH="amd64"
+fi
+
+docker run --rm "debify:$VERSION-$ARCH" config script > docker-debify
 chmod +x docker-debify
-DEBIFY_IMAGE=debify:$VERSION DEBIFY_ENTRYPOINT=ci/test.sh ./docker-debify
+DEBIFY_IMAGE=debify:$VERSION-$ARCH DEBIFY_ENTRYPOINT=ci/test.sh ./docker-debify